refactor: split security command family
